correccion de vista en productos y se agrego estructura de archivos de los users
