+++ src/controllers/mainController.js
@@ -16,20 +16,7 @@ const controller = {
   register: (req, res) => {
     res.render('./auth/register');
   },
-  createUser:(req,res)=> {
-    let user ={
-      id: users.length + 1,
-      nombre: req.body.nombre + " " + req.body.apellido,
-      ubicacion: req.body.ubicacion,
-      imagen: "",
-      email: req.body.email,
-      contraseña: req.body.contraseña,
-    }
-   users.push(user);
-   let usersJSON =JSON.stringify(users);
-   fs.writeFileSync(usersFilePath,usersJSON);
-   res.redirect("/")
-  }
+  
 };
 
 module.exports = controller;
