+++ src/routes/users.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const usersController = require("../controllers/usersController");
+const router = express.Router();
+
+router.get("/",usersController.index)
+
+module.exports = router
