+++ src/views/products/index.ejs
@@ -15,7 +15,7 @@
             <a href="/products/<%= producto.id %>/detail/" >
                 <article>
                     <div class="image" style="background-image: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)), url('/uploads/<%= producto.imagen %>');">
-                        <h2><%= producto.naombre %></h2>
+                        <h2><%= producto.nombre %></h2>
                         <p><%= producto.ubicacion %></p>
                     </div>
                     <p><%= producto.deporte %></p>
