+++ src/views/users/index.ejs
@@ -1,0 +1,1 @@
+hola
